# erp-ai/src/main/java/com/own/erp/ai/response/AiChatSessionResponse.java
package com.own.erp.ai.response;

import com.own.erp.ai.entity.AiChatSession;
import lombok.Builder;

import java.time.LocalDateTime;

/**
 * @author : chenyi
 * @Date : 2026/9/6
 * @Description : AI会话对外结构(docs/07 §1:XxxResponse 只声明允许对外的字段)
 *     record+@Builder(模型可变性分级 docs/07 §1):读侧不可变;from 用 builder 命名传参防相邻同类型字段错位
 *     生成器默认全字段映射:敏感/服务端管理字段必须人工删减,不该出去的字段不建进组件即编译期封死
 */
@Builder
public record AiChatSessionResponse(

        /** 主键 */
        Long id,

        /** 所属用户ID(sys_user.id,会话归属校验依据) */
        Long userId,

        /** 会话标题(首条提问截断) */
        String title,

        /** 创建时间 */
        LocalDateTime createdAt,

        /** 更新时间 */
        LocalDateTime updatedAt

) {

    /** 实体 → Response 显式逐字段映射(禁反射拷贝,漏字段编译期可见);builder 命名传参防相邻同类型字段错位 */
    public static AiChatSessionResponse from(AiChatSession entity) {
        return AiChatSessionResponse.builder()
                .id(entity.getId())
                .userId(entity.getUserId())
                .title(entity.getTitle())
                .createdAt(entity.getCreatedAt())
                .updatedAt(entity.getUpdatedAt())
                .build();
    }
}