# erp-web/tools/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8088",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/ai/suggestions/{id}/ignore": {
      "post": {
        "tags": [
          "AI建议"
        ],
        "summary": "忽略建议",
        "description": "0待确认→2已忽略,条件更新即守卫,与采纳终态互斥",
        "operationId": "ignore",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/suggestions/{id}/adopt": {
      "post": {
        "tags": [
          "AI建议"
        ],
        "summary": "采纳建议",
        "description": "0待确认→1已采纳,条件更新即守卫(cas 脱靶=已处理/不存在,报错);仅落确认状态,业务动作由人工走对应业务接口",
        "operationId": "adopt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ai/chat/sessions": {
      "get": {
        "tags": [
          "AI助手"
        ],
        "summary": "我的会话分页",
        "description": "按最近更新倒序;仅本人会话(归属服务端强制)",
        "operationId": "sessions",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AiChatSessionQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageAiChatSessionResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "新建会话",
        "description": "标题可空,默认'新会话',首条消息后自动回填为消息摘要",
        "operationId": "createSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/chat/sessions/{id}/chat": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "流式对话(SSE)",
        "description": "text/event-stream,逐段返回 AI 回复;消息内容取 body.message",
        "operationId": "chat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSendCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/chat/sessions/{id}/chat-sync": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "同步对话",
        "description": "一次性返回完整回复(带工具查询,耗时高于流式首包)",
        "operationId": "chatSync",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSendCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ai/suggestions": {
      "get": {
        "tags": [
          "AI建议"
        ],
        "summary": "分页查询AI建议",
        "operationId": "page_18",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AiSuggestionQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageAiSuggestionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/suggestions/{id}": {
      "get": {
        "tags": [
          "AI建议"
        ],
        "summary": "AI建议详情",
        "description": "不存在返回 null data",
        "operationId": "get_14",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAiSuggestionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/chat/sessions/{id}/messages": {
      "get": {
        "tags": [
          "AI助手"
        ],
        "summary": "会话历史消息",
        "description": "时间正序;仅本人会话可查,越权/不存在统一报'会话不存在'",
        "operationId": "messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListAiChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {}
        }
      },
...
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "SessionCreateCommand": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
      "ChatSendCommand": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "message"
        ]
      },
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "asc": {
            "type": "boolean"
          }
        }
      },
...
      "AiSuggestionQuery": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "shopId": {
            "type": "integer",
            "format": "int64"
          },
          "skuId": {
            "type": "integer",
            "format": "int64"
          },
          "suggestionType": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "AiSuggestionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "suggestionType": {
            "type": "string"
          },
          "shopId": {
            "type": "integer",
            "format": "int64"
          },
          "skuId": {
            "type": "integer",
            "format": "int64"
          },
          "refType": {
            "type": "string"
          },
          "refId": {
            "type": "integer",
            "format": "int64"
          },
          "payloadJson": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "riskLevel": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "confirmedBy": {
            "type": "integer",
            "format": "int64"
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageAiSuggestionResponse": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiSuggestionResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "writeOnly": true
          },
          "optimizeCountSql": {
            "$ref": "#/components/schemas/PageAiSuggestionResponse",
            "writeOnly": true
          },
          "searchCount": {
            "$ref": "#/components/schemas/PageAiSuggestionResponse",
            "writeOnly": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "writeOnly": true
          },
          "maxLimit": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "countId": {
            "type": "string",
            "writeOnly": true
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        }
      },
      "ResultPageAiSuggestionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageAiSuggestionResponse"
          }
        }
      },
      "ResultAiSuggestionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AiSuggestionResponse"
          }
        }
      },
      "AiChatSessionQuery": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AiChatSessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageAiChatSessionResponse": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiChatSessionResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "writeOnly": true
          },
          "optimizeCountSql": {
            "$ref": "#/components/schemas/PageAiChatSessionResponse",
            "writeOnly": true
          },
          "searchCount": {
            "$ref": "#/components/schemas/PageAiChatSessionResponse",
            "writeOnly": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "writeOnly": true
          },
          "maxLimit": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "countId": {
            "type": "string",
            "writeOnly": true
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        }
      },
      "ResultPageAiChatSessionResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageAiChatSessionResponse"
          }
        }
      },
      "AiChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sessionId": {
            "type": "integer",
            "format": "int64"
          },
          "role": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "toolName": {
            "type": "string"
          },
          "promptTokens": {
            "type": "integer",
            "format": "int32"
          },
          "completionTokens": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultListAiChatMessageResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiChatMessageResponse"
            }
          }
        }
      },
...
    }
  }
}